requirements.txt
pydub~=0.25.1
aiogram~=3.20.0.post0
utils~=1.0.2
pandas~=2.2.3
google~=3.0.0
protobuf~=5.29.4
aiogram==3.2.0
pydub
google-cloud-speech
openai
python-dotenv

